requirements.txt
# requirements.txt — Perimeter Defense Detection System
# Install: pip install -r requirements.txt

# Core
ultralytics>=8.0.0          # YOLOv8 (includes torch dependency management)
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
gradio
torch
torchvision
huggingface_hub